Automatically generate Visual Studio project/solution files. * Now using premake to generate project files. * New files are now placed under the build folder that is not under source control and that will contain all the intermediate/output files. * Old build folders and projects files removed from source control and ignore list. Now is a good time to start working with a fresh checkout. * Default save game folder is now changed to be sibling of the rom folder. The folder will be created if it doesn't exist. * Added support for 64-bit build targets. * Fixed all warnings related to 32/64 bit builds.

+++ ManyNES/APU.cpp
@@ -321,9 +321,9 @@ namespace NES
     void APU::setSoundBuffer(int16_t* buffer, size_t size)
     {
         mSoundBuffer = buffer;
-        mSoundBufferSize = size;
+        mSoundBufferSize = static_cast<uint32_t>(size);
         mSoundBufferOffset = 0;
-        mSampleSpeed = mMasterClockFrequency / (60 * size);
+        mSampleSpeed = mMasterClockFrequency / (60 * static_cast<uint32_t>(size));
     }
 
     void APU::updateEnvelopesAndLinearCounter()
