+++ ManyNES/GameSession.h
@@ -18,7 +18,7 @@ public:
     bool saveGameState();
     void serializeGameState(NES::ISerializer& serializer);
     void setRenderBuffer(void* buffer, uint32_t pitch);
-    void setSoundBuffer(void* buffer, uint32_t size);
+    void setSoundBuffer(void* buffer, size_t size);
     void setController(uint32_t index, uint8_t value);
     void execute();
